Provided manual updating steps for Cloudflared
Changed documentation to first list manual method to
update the cloudflared binary, and then have a subsequent
paragraph on automatic updating if desired by the user.

Also, while small, added a note that cloudflared will
work with other DoH providers (was useful when there was
a cloudflare outage in my region).