Update the unspecified IP blocking documentation
Updated the documentation reflecting 0.0.0.0 being treated like 127.0.0.1 in some cases. This behavior can trigger in a bunch of networking stacks, e.g. Zoom on macOS will try to connect to a local webserver if some zoom urls are blocked. Similarly, most of Synology's software running on Linux will try to do the same.

Signed-off-by: Vladimir Pouzanov <farcaller@gmail.com>